+++ urlfix/urlfix.py
@@ -6,7 +6,6 @@ def find_links(input_string):
     """
 
     :param input_string String to search for markdown like links
-    :param input_format of the links. Currently supports one of md or txt
     :return  Extracts links from MarkDown code
 
     """
@@ -15,25 +14,28 @@ def find_links(input_string):
     return matched_urls
 
 
-
-
 def show_parsed_urls(input_file):
+    """
 
+    :param input_file: File whose URLs are to be parsed
+    :return: URLs in the target file
+    """
     input_file = open(input_file)
-    matched_urls = list(filter(None,[re.sub("\n|\!", "", find_links(line)) for line in input_file]))
+    matched_urls = list(filter(None, [re.sub("\n|\!", "", find_links(line)) for line in input_file]))
     input_file.close()
     return matched_urls
-
-
 
 
 # Keep url, replacement as pair and replace downstream?
 # Directly replace url in document?
 
-def visit_urls(input_file, verbose=False, return_matched=False):
+def replace_urls(input_file, output_file="replacement.txt", inplace=False, verbose=False, return_matched=False):
     """
 
     :param input_file File whose links you need to replace or check.
+
+    :param output_file File to write to.
+    :param inplace Logical. Determines if you need to replace URLs inplace. Defaults to False.
 
     :param verbose Logical. Should you be notified of what URLs have moved? Defaults to False.
 
@@ -43,7 +45,9 @@ def visit_urls(input_file, verbose=False, return_matched=False):
 
 
     """
-    # TODO: Finalize string replacement in input file.
+
+    # TODO: Speed up file writing and URL checks, avoid unnecessary loops
+    # TODO: Work with several files in a directory
 
     matched_urls = show_parsed_urls(input_file)
     number_moved = 0
@@ -52,12 +56,16 @@ def visit_urls(input_file, verbose=False, return_matched=False):
         url_used = visited.geturl()
 
         if url_used != target_url:
-            if verbose:
-                print(f"{target_url} moved to {url_used}")
+            if inplace:
+                output_file = input_file
+            with open(input_file, "r") as input_f, open(output_file, "w") as out_f:
+                for line in input_f:
+                    if verbose:
+                        print(f"{target_url} replaced with {url_used} in {out_f}")
+                    out_f.write(line.replace(target_url, url_used))
 
             matched_urls[index] = url_used
             number_moved += 1
     information = "URLS have changed" if number_moved != 1 else "URL has changed"
     print(f"{number_moved} {information}")
     return matched_urls if not return_matched else matched_urls, number_moved
-