+++ changelog.md
@@ -1,8 +1,11 @@
 # Welcome to urlfix's changelog 
 
 
-**urlfix 0.1.0**
+**urlfix 0.1.1**
 
+* `visit_urls` was renamed to `replace_urls` and extended to allow inplace replacement (or not)
+ as well as writing to an output file.
+  
 * `fixurls` was renamed to `urlfix`.
 
 * `check_url` was removed but may be replaced. 
@@ -15,6 +18,8 @@
 
 * Fixed issues with installation
 
+
+**urlfix 0.1.0**
 
 * Initial release to preserve name on PyPI.