+++ docs/source/changelog.rst
@@ -2,8 +2,12 @@
 Welcome to urlfix's changelog
 =============================
 
-**urlfix 0.1.0**
+**urlfix 0.1.1**
 
+
+* 
+  ``visit_urls`` was renamed to ``replace_urls`` and extended to allow inplace replacement (or not)
+  as well as writing to an output file.
 
 * 
   ``fixurls`` was renamed to ``urlfix``.
@@ -22,6 +26,8 @@ Welcome to urlfix's changelog
 
 * 
   Fixed issues with installation
+
+**urlfix 0.1.0**
 
 
 * Initial release to preserve name on PyPI.